+++ lib/install/install.rb
@@ -1,0 +1,2 @@
+say "Copy l10n.js to app/javascript"
+copy_file "#{__dir__}/l10n.js", "app/javascript/l10n.js"