+++ lib/l10n/engine.rb
@@ -1,7 +1,0 @@
-require 'rails/engine'
-
-module L10n
-  class Engine < ::Rails::Engine
-    
-  end
-end